requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiosqlite
colorlog
colorama
PyJWT
python-dotenv